+++ src/test/java/ru/yandex/money/common/dbqueue/spring/wiring/SpringLifecycleConfiguration.java
@@ -212,7 +212,7 @@ public class SpringLifecycleConfiguration {
         }
 
         @Override
-        public void crashedPickTask(@Nonnull QueueShardId shardId, @Nonnull QueueLocation location, @Nonnull Throwable exc) {
+        public void crashed(@Nonnull QueueShardId shardId, @Nonnull QueueLocation location, @Nonnull Throwable exc) {
             EVENTS.add("queue crashed");
         }
     }