+++ CHANGELOG.md
@@ -12,6 +12,11 @@ Given version x.y.z
 
 **In case of x == 0 backward compatibility is not guaranteed.**
 
+### [0.0.9-SNAPSHOT]() (26-08-2017)
+
+* Fixed ability to disable queue processing via thread count
+* Renamed ThreadLifecycleListener#crashedOnPickedTask to ThreadLifecycleListener#crashed
+
 ### [0.0.8]() (24-08-2017)
 
 Renamed classes: